# gin-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golang 微服務範例\n更多說明參考 https://github.com/Klay24-Huang/go-grpc-pratice",
        "title": "Go gRPC simple demo",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/orders": {
            "get": {
                "tags": [
                    "order"
                ],
                "summary": "取得order列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/order.OrderResponse"
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "order"
                ],
                "summary": "更新order資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "order info",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.OrderReply"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "order"
                ],
                "summary": "新增order",
                "parameters": [
                    {
                        "description": "order info",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.OrderReply"
                        }
                    }
                }
            }
        },
        "/orders/{order_id}": {
            "get": {
                "tags": [
                    "order"
                ],
                "summary": "取得指定id order資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "order id",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order.OrderResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "order"
                ],
                "summary": "刪除指定id order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "order id",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "取得product列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/product.ProductResponse"
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "product"
                ],
                "summary": "更新product資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "product info",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.ProductReply"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "product"
                ],
                "summary": "新增product",
                "parameters": [
                    {
                        "description": "product info",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.ProductReply"
                        }
                    }
                }
            }
        },
        "/products/{product_id}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "取得指定id product資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "product id",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product.ProductResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "取得user列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user.UserResponse"
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "更新user資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_service.UserReply"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "新增user",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_service.UserReply"
                        }
                    }
                }
            }
        },
        "/users/{user_id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "取得指定id user資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UserResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "__.OrderProductReply": {
            "type": "object",
            "properties": {
                "productAmount": {
                    "type": "integer"
                },
                "productId": {
                    "type": "integer"
                }
            }
        },
        "__.OrderReply": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/__.OrderProductReply"
                    }
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "__.ProductReply": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "order.OrderProduct": {
            "type": "object",
            "required": [
                "product_amount",
                "product_id"
            ],
            "properties": {
                "product_amount": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "order.OrderProductsResponse": {
            "type": "object",
            "properties": {
                "product_amount": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "product_price": {
                    "type": "integer"
                }
            }
        },
        "order.OrderRequest": {
            "type": "object",
            "required": [
                "products",
                "user_id"
            ],
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/order.OrderProduct"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "order.OrderResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "order_price": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/order.OrderProductsResponse"
                    }
                },
                "user": {
                    "$ref": "#/definitions/order.OrderUserResponse"
                }
            }
        },
        "order.OrderUserResponse": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "product.ProductRequest": {
            "type": "object",
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "product.ProductResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "user.UserRequest": {
            "type": "object",
            "required": [
                "account",
                "name",
                "phone"
            ],
            "properties": {
                "account": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "user.UserResponse": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "user_service.UserReply": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "id": {
                    "description": "UserInfo userInfo = 1;\nErrorMessage error = 2;",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        }
    }
}